% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{checkboxInput}
\alias{checkboxInput}
\title{Create an advanced checkbox input}
\usage{
checkboxInput(inputId, label, value = FALSE, width = NULL, inline = FALSE)
}
\arguments{
\item{inputId}{The identification name}

\item{label}{The label for the input}

\item{value}{The current value of the input}

\item{width}{width of the text input}

\item{inline}{make inline or block format}
}
\value{
HTML elements of a checkbox
}
\description{
Modifications to checkboxinput to allow added styles
}
\note{
For more information on the features of the form, visit the examples section of the help documentation
}
\examples{
        checkboxInput("somevalue", "Some value", FALSE)
        checkboxInput("somevalue", "Some value", value = FALSE, inline = TRUE)
        checkboxInput("somevalue", "Some value", FALSE)


}
