% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cssjsinclude}
\alias{cssjsinclude}
\title{Include stylesheets and scripts}
\usage{
cssjsinclude(template, color)
}
\arguments{
\item{template}{The template type}

\item{color}{The numeric style of template}
}
\value{
A list of files to be inserted in the header of a page
}
\description{
Use the package scripts and stylesheets in a page
}
\examples{
cssjsinclude('color','3')
}
