% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{dashboardBody}
\alias{dashboardBody}
\title{Create the body section of the application}
\usage{
dashboardBody(header, ...)
}
\arguments{
\item{header}{OPTIONAL. Items to display in the header section (use the titlePanel() function to set this property).}

\item{...}{The elements to include within the body of the modal}
}
\value{
An HTML of the body of the page
}
\description{
Create a simple body containing a header and a content for the main body
}
\note{
Endeavor to use as standalone and not within the fluidPage, as this function it already called within fluidPage
}
\examples{

if (interactive()) {
dashboardBody(
  header = titlePanel(
    left = "Sample nextGenShinyApps Title",
    right = shiny::icon("user")
  ),
  "sample text for main body"
)
}

}
