% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{flexBox}
\alias{flexBox}
\title{Container to fit all containers horizontally}
\usage{
flexBox(..., border = FALSE, shadow = FALSE)
}
\arguments{
\item{...}{List of content}

\item{border}{Should border be declared for the panel}

\item{shadow}{Should a shadow be added to the panel}
}
\value{
Creates a a box to fit all boxes within
}
\description{
Fit containers horizontally
}
\examples{
\donttest{
 flexBox(shiny::div("Div 1"),shiny::div("Div 2"),shiny::div("Div 1"))
 }
}
