% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleloader.R
\name{load.example}
\alias{load.example}
\title{Load examples for the package}
\usage{
load.example(
  example = c("Plain", "noSideBar", "Card", "Tab", "Modal", "Form", "Button",
    "Spinner", "Alert", "Accordion")
)
}
\arguments{
\item{example}{choose the example to show - "Plain","noSideBar","Card","Tab","Modal","Form","Button","Spinner","Alert","Accordion"}
}
\value{
A rendered HTML of the user specified example file
}
\description{
Example shiny applications spanning aspects of the package
}
\examples{
\donttest{
if (interactive()) {
load.example(example = "Card")
}
}

}
