% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{mainPanel}
\alias{mainPanel}
\title{Main panel to display content}
\usage{
mainPanel(..., width = 8, border = FALSE, shadow = FALSE)
}
\arguments{
\item{...}{List of content}

\item{width}{Width of the main panel}

\item{border}{Should border be declared for the panel}

\item{shadow}{Should a shadow be added to the panel}
}
\value{
Creates a container for displaying contents
}
\description{
Customizable main panel for inclusion of various UI elements
}
\note{
For more information on the features of the main panel, look through the Github examples
}
\examples{
\donttest{
 mainPanel('content 1')
 }
}
