% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modal.header}
\alias{modal.header}
\title{Create hyperlink modal section that appears ONLY when the sidebar logo is clicked}
\usage{
modal.header(...)
}
\arguments{
\item{...}{The list of tabs to include}
}
\value{
An HTML containing elements of links to be inserted in the header of a page
}
\description{
Additional tab section for additional links
}
\examples{
if (interactive()) {
list1 <- list(
  title = "Home", icon = shiny::icon("home"),
  link = "https://google.com"
)
list2 <- list(
  title = "Docs", icon = shiny::icon("folder"),
  link = "https://obi.obianom.com"
)

modal.header(list(list1, list2))
}

}
