% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{nav}
\alias{nav}
\title{Nav tag}
\usage{
nav(class, id = NULL, role = NULL, ...)
}
\arguments{
\item{class}{The class of the navigation container}

\item{id}{The identification of the navigation container}

\item{role}{The character role of the container on the page}

\item{...}{The content of the container}
}
\value{
HTML content of a container with type nav
}
\description{
A nav tag for creating HTML navigations
}
\examples{
nav('sample','id1','sample','some content')
}
