% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{paginatorBtn}
\alias{paginatorBtn}
\title{Create a pagination set}
\usage{
paginatorBtn(inputId, ..., active, buttons, nav.type = c("text", "icons"))
}
\arguments{
\item{inputId}{input id}

\item{...}{Other inputs}

\item{active}{which is the active button}

\item{buttons}{what buttons to include between the previous and next}

\item{nav.type}{navigation icon types e.g 'text' or 'icons'}
}
\value{
HTML of the submit buttons to insert into a page
}
\description{
Pagination buttons and arrows
}
\examples{
if (interactive()) {
paginatorBtn("sampleId"
)
}
}
