% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rand.num}
\alias{rand.num}
\title{Random number betwen 1 and 10000}
\usage{
rand.num(num)
}
\arguments{
\item{num}{The number of numbers to return}
}
\value{
One or more numbers
}
\description{
One or more random numbers
}
\examples{
rand.num(10)
}
