% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{row}
\alias{row}
\title{Generate a row div}
\usage{
row(...)
}
\arguments{
\item{...}{The elements to include within the body of the row}
}
\value{
An HTML containing elements of a container with class row to be embedded in a page
}
\description{
A simple row div
}
\examples{
row(shiny::div(width=12,"Hello nextGenShinyApps"))
}
