% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{setup.toolbar.buttons}
\alias{setup.toolbar.buttons}
\title{Generate toolbar buttons}
\usage{
setup.toolbar.buttons(...)
}
\arguments{
\item{...}{The list of buttons to display}
}
\value{
HTML code of a container containing items to be inserted in the toolbar
}
\description{
Use within a card to display toolbar
}
\note{
For more information on the features of a toolbar within a card, visit the examples section of the help documentation
}
\examples{
\donttest{
 setup.toolbar.buttons(list(maximize=TRUE,collapse=TRUE,close=TRUE))
 }
}
