% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{setup.toolbar.menu}
\alias{setup.toolbar.menu}
\title{Generate toolbar menu}
\usage{
setup.toolbar.menu(...)
}
\arguments{
\item{...}{The list declaring whether to show menu}
}
\value{
HTML code of a container containing menu to be inserted in the toolbar if declared TRUE
}
\description{
Use within a card to display menu
}
\note{
For more information on the features of a toolbar within a card, visit the examples section of the help documentation
}
\examples{
\donttest{
 setup.toolbar.menu(list(menu=TRUE))
 }
}
