% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sidebarLayout}
\alias{sidebarLayout}
\title{Sidebar layout}
\usage{
sidebarLayout(...)
}
\arguments{
\item{...}{content of the sidebar layout}
}
\value{
An HTML containing elements in a container
}
\description{
Container for sidebar, optional
}
\examples{
sidebarLayout("sample text")
}
