% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R, R/sidebar.R
\name{sidebarPanel}
\alias{sidebarPanel}
\title{New sidebar panel to display content}
\usage{
sidebarPanel(..., title = "TitleApp", footer = FALSE)

sidebarPanel(..., title = "TitleApp", footer = FALSE)
}
\arguments{
\item{...}{The elements to include within the body of the sidebar}

\item{title}{Title of the sidebar content}

\item{footer}{Whether or not to enable the footer content}

\item{width}{Width of the sidebar panel}

\item{border}{Should border be declared for the panel}

\item{shadow}{Should a shadow be added to the panel}
}
\value{
Creates a container for displaying contents

An HTML containing elements of a sidebar to be embedded in a page
}
\description{
Customizable sidebar panel for inclusion of various UI elements

Creates an advanced sidebar panel with colors corresponding to a chosen theme style
}
\note{
For more information on the features of the sidebar panel, look through the Github examples

This global layout can be applied to a variety of shiny app and dashboard, with or without a sidebar. See the example below.
}
\examples{
\donttest{
 sidebarPanel('content 2')
 }
\donttest{
if (interactive()) {
  library(shiny)
  library(nextGenShinyApps)

  shiny::shinyApp(
    ui = fluidPage(
      style = "7",
      header = NULL,
      sidebar = sidebarPanel(
        title = "myApp",
        "Sample sidebar contents",
        footer = FALSE
      ),
      "Plain content"
    ),
    server = function(input, output) {}
  )
}
}
}
