% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{sortablegrid}
\alias{sortablegrid}
\title{Generate a sortable grid}
\usage{
sortablegrid(..., width = 6, id = NULL)
}
\arguments{
\item{...}{The elements to include within the body of the grid}

\item{width}{The width of the grid}

\item{id}{unique id of grid}
}
\value{
HTML code of a container that allows items within it to be draggable
}
\description{
A grid that holds draggable items
}
\note{
For more information on the features of a sortable grid, visit the examples section of the help documentation
}
\examples{
 sortablegrid("item1",width=12)
}
