% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{tabPanel}
\alias{tabPanel}
\title{Create a tab panel item}
\usage{
tabPanel(title, ...)
}
\arguments{
\item{title}{title of the tab}

\item{...}{content of the tab}
}
\value{
An list containing the title and content of a tab
}
\description{
Create a tab panel item that is enclosed by a tabsetPanel
}
\examples{
if (interactive()) {
tabPanel("Summary", "Convallis aesus.")
tabPanel("Summary", "nextGenShinyAppss.")
}

}
