% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{template.loc}
\alias{template.loc}
\title{Template location full text}
\usage{
template.loc(template = "core")
}
\arguments{
\item{template}{The type of template to fetch}
}
\value{
A path for the location of the package
}
\description{
Fetch the location of the scripts
}
\examples{
template.loc('core')
}
