% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{titlePanel}
\alias{titlePanel}
\title{Title panel for the header of the application}
\usage{
titlePanel(left = "Sample Tile", right = NULL, link = "#")
}
\arguments{
\item{left}{The title text for the header}

\item{right}{Content to include on the top right corner}

\item{link}{Hyperlink to navigate when clicked}
}
\value{
An HTML containing elements to insert in a title
}
\description{
Used to embed the header within the body of the application
}
\examples{
if (interactive()) {
  titlePanel(
    left = "Sample App Title",
    right = shiny::div("Image/logo", shiny::icon("trash"))
  )
  }

}
