% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{wrapper}
\alias{wrapper}
\title{Panel wrapper}
\usage{
wrapper(...)
}
\arguments{
\item{...}{content of the container}
}
\value{
Container containing various panels
}
\description{
Wrapper to hold all panels
}
\examples{
if (interactive()) {
wrapper(
  altPanel("Sample left content"),
  mainPanel("Sample right content")
)
}

}
