% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{paginatorBtn}
\alias{paginatorBtn}
\title{Create a pagination set}
\usage{
paginatorBtn(inputId, n, active, buttons, nav.type = c("text", "icons"))
}
\arguments{
\item{inputId}{input id}

\item{n}{Number of pages}

\item{active}{which is the active button}

\item{buttons}{what buttons to include between the previous and next}

\item{nav.type}{navigation icon types e.g 'text' or 'icons'}
}
\value{
HTML of the submit buttons to insert into a page
}
\description{
Pagination buttons and arrows as of v1.7
}
\examples{
if (interactive()) {
paginatorBtn("sampleId"
)
}
}
